sage-ai-backend/simple_agent_requirements.txt
"# Simple LiveKit Agents Requirements" 

# Simple LiveKit Agents Requirements - Context7 MCP Recommended Approach
# This replaces our complex 15+ dependency approach with just 2 core packages!

# Core LiveKit Agents with OpenAI Realtime API and turn detection
livekit-agents[openai,turn-detector]>=1.0

# Environment variable management  
python-dotenv>=1.0

# That's it! No complex audio processing, no custom bridges, no manual room management!
# OpenAI Realtime API + LiveKit Agents handles everything automatically. 
